% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWORDmethods.R
\name{extractCL}
\alias{extractCL}
\title{Extract node information from the SWORD database (Alteneau et al., 2021) for multiple reaches and reorders the nodes from downstrean to upstream if these are ordered wrongly
Important the function requires the SWORD database! If you do not have this please download from https://www.swordexplorer.com}
\usage{
extractCL(myreaches, path2sword, SWORDv, fix = FALSE, addChain = FALSE)
}
\arguments{
\item{path2sword}{Path to the location of the SWORD database}

\item{SWORDv}{the version of the SWORD database}

\item{thisreach}{Reach id from the SWORD database}
}
\description{
Extract node information from the SWORD database (Alteneau et al., 2021) for multiple reaches and reorders the nodes from downstrean to upstream if these are ordered wrongly
Important the function requires the SWORD database! If you do not have this please download from https://www.swordexplorer.com
}
\examples{
path2sword<-"/home/karina/SWORD/SWORD_v15_nc/netcdf"
SWORDv='15'
myreaches<-c(21406100041,21406100031,21406100021)
out<-extractCL(myreaches,path2sword,SWORDv)
}
