% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARRDmethods.R
\name{getAltReaches}
\alias{getAltReaches}
\title{Extract nadir altimetry data from the Altimetry River Reach database (ARRD). This function requires
access to a basin sqlite file. Basin sqlite files can be acessed from ...}
\usage{
getAltReaches(dbfile, mybasin, myreaches)
}
\arguments{
\item{dbfile}{a particular basin sqlite file e.g. "7428.sqlite". The basin number is the HydroShed level 4 PFAF_ID}

\item{mybasin}{HydroShed basin PFAF_ID (level 4)}

\item{myReaches}{a vector with SWORD reach numbers}
}
\description{
Extract nadir altimetry data from the Altimetry River Reach database (ARRD). This function requires
access to a basin sqlite file. Basin sqlite files can be acessed from ...
}
\examples{
dbfile<-"7428.sqlite"
mybasin<-'7428'
myreaches<-c(74287400183, 74287400163, 74287400151, 74287400173)
out<-getAltReaches(dbfile,mybasin,myreaches)
}
