% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWORDmethods.R
\name{getReachInfo}
\alias{getReachInfo}
\title{Extract node information from the SWORD database (Alteneau et al., 2021) for a specific reach
Important the function requires the SWORD database! If you do not have this please download from https://www.swordexplorer.com}
\usage{
getReachInfo(thisreach, path2sword, SWORDv)
}
\arguments{
\item{thisreach}{Reach id from the SWORD database}

\item{path2sword}{Path to the location of the SWORD database}

\item{SWORDv}{the version of the SWORD database}
}
\description{
Extract node information from the SWORD database (Alteneau et al., 2021) for a specific reach
Important the function requires the SWORD database! If you do not have this please download from https://www.swordexplorer.com
}
\examples{
path2sword<-"/home/karina/SWORD/SWORD_v15_nc/netcdf"
SWORDv='15'
thisreach<-"21406100031"
out<-getReachInfo(thisreachreach,path2sword,SWORDv)
}
