% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverReachTS.R
\name{plot.ARRM}
\alias{plot.ARRM}
\title{Plot an object returned by the function riverReachTS()}
\usage{
\method{plot}{ARRM}(fit, dat = NULL, doSave = FALSE, type = "F", ...)
}
\arguments{
\item{fit}{Object returned by get.TS()}

\item{dat}{The raw water level data can be added to the plot.}

\item{...}{Additional argumants to plot}
}
\description{
Plot an object returned by the function riverReachTS()
}
\keyword{plot}
