% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverReachTS.R
\name{riverReachTS}
\alias{riverReachTS}
\title{Fit riverReachTS}
\usage{
riverReachTS(
  dat,
  VS = NULL,
  tdim = NULL,
  xdim = 20,
  Nknot = 10,
  nit = 5,
  cutp = 0.05
)
}
\arguments{
\item{dat}{Data}

\item{VS}{Reach distance of virtual stations}

\item{tdim}{Number of time steps in the model}

\item{xdim}{Number of space steps in the model}

\item{Nknot}{Number of spline nodes to describe the mean river reach profile}

\item{nit}{Number of model iterations, usually 2-5 is enough}

\item{cutp}{The fraction of expected outliers}
}
\description{
Fit riverReachTS
}
\examples{
fit<-riverReachTS(dat)
}
